+++ MinkowskiEngine/__init__.py
@@ -32,13 +32,15 @@ sys.path.append(file_dir)
 # Must be imported first to load all required shared libs
 import torch
 
+from MinkowskiEngineBackend import MemoryManagerBackend
+
 from SparseTensor import SparseTensor, SparseTensorOperationMode, SparseTensorQuantizationMode, \
     set_sparse_tensor_operation_mode, sparse_tensor_operation_mode, clear_global_coords_man
 
 from Common import RegionType, convert_to_int_tensor, convert_region_type, \
     MinkowskiModuleBase, KernelGenerator, GlobalPoolingMode
 
-from MinkowskiCoords import CoordsKey, CoordsManager
+from MinkowskiCoords import CoordsKey, CoordsManager, set_memory_manager_backend
 
 from MinkowskiConvolution import MinkowskiConvolutionFunction, MinkowskiConvolution, \
     MinkowskiConvolutionTransposeFunction, MinkowskiConvolutionTranspose