+++ tests/coords.py
@@ -26,7 +26,8 @@ import unittest
 import torch
 import numpy as np
 
-from MinkowskiEngine import CoordsKey, CoordsManager
+import MinkowskiEngine as ME
+from MinkowskiEngine import CoordsKey, CoordsManager, MemoryManagerBackend
 
 from tests.common import data_loader
 
@@ -166,6 +167,13 @@ class Test(unittest.TestCase):
         cm.create_coords_key(coords)
 
         self.assertTrue(cm.get_batch_size() == 2)
+
+    def test_memory_manager_backend(self):
+        CoordsManager(memory_manager_backend=MemoryManagerBackend.CUDA, D=2)
+        CoordsManager(memory_manager_backend=MemoryManagerBackend.PYTORCH, D=2)
+
+        ME.set_memory_manager_backend(MemoryManagerBackend.PYTORCH)
+        CoordsManager(D=2)
 
 
 if __name__ == '__main__':